Monotone-Parent: 0b791018ccfc1dc1f549e84e51f34546204f30f0 Monotone-Revision: 194f26d55e34840849843f8a494a9c59b251e3af
Monotone-Author: wsourdeau@inverse.ca
Monotone-Date: 2010-03-24T15:23:59
Monotone-Branch: ca.inverse.calendar

+++ src/calendar/base/content/calendar-unifinder.js
@@ -178,14 +178,14 @@ var unifinderObserver = {
 
     // calICompositeObserver:
     onCalendarAdded: function uO_onCalendarAdded(aAddedCalendar) {
-        if (!aAddedCalendar.getProperty("disabled")) {
+        if (!isUnifinderHidden() && !aAddedCalendar.getProperty("disabled")) {
             addItemsFromCalendar(aAddedCalendar,
                                  addItemsFromSingleCalendarInternal);
         }
     },
 
     onCalendarRemoved: function uO_onCalendarRemoved(aDeletedCalendar) {
-        if (!aDeletedCalendar.getProperty("disabled")) {
+        if (!isUnifinderHidden() && !aDeletedCalendar.getProperty("disabled")) {
             deleteItemsFromCalendar(aDeletedCalendar);
         }
     },
@@ -890,8 +890,8 @@ function addItemsFromCalendar(aCalendar, aAddItemsInternalFunc) {
 function deleteItemsFromCalendar(aCalendar) {
     var items = [];
 
-    for (var i = unifinderTreeView.eventArray.length; i > 0; i--) {
-        var item = unifinderTreeView.eventArray[i-1];
+    for (var i = 0; i < unifinderTreeView.eventArray.length; i++) {
+        var item = unifinderTreeView.eventArray[i];
         if (item.calendar.id == aCalendar.id)
             items.push(item);
     }
